requirements.txt
fastapi
uvicorn
pydantic>=2.0
openai
httpx
python-dotenv
